DPC-567: Refactored Organization creation command.
Org registration is now part of the organization command tree.
