+++ dpc-api/src/main/java/gov/cms/dpc/api/cli/organizations/OrganizationRegistration.java
@@ -1,16 +1,13 @@
-package gov.cms.dpc.api.cli;
+package gov.cms.dpc.api.cli.organizations;
 
-import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.parser.IParser;
 import ca.uhn.fhir.rest.client.api.IGenericClient;
-import ca.uhn.fhir.rest.client.api.ServerValidationModeEnum;
+import gov.cms.dpc.api.cli.AbstractAttributionCommand;
 import gov.cms.dpc.fhir.FHIRMediaTypes;
-import io.dropwizard.cli.Command;
 import io.dropwizard.setup.Bootstrap;
 import net.sourceforge.argparse4j.inf.Namespace;
 import net.sourceforge.argparse4j.inf.Subparser;
 import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
@@ -18,8 +15,6 @@ import org.apache.http.util.EntityUtils;
 import org.hl7.fhir.dstu3.model.Bundle;
 import org.hl7.fhir.dstu3.model.Organization;
 import org.hl7.fhir.dstu3.model.Parameters;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -28,35 +23,22 @@ import java.nio.file.FileSystems;
 import java.nio.file.Path;
 import java.util.UUID;
 
-public class OrgRegistrationCommand extends Command {
+public class OrganizationRegistration extends AbstractAttributionCommand {
 
     private static final String ORG_FILE = "org-file";
-    private static final String ORG_ID = "org-id";
-    private static final String ATTR_HOSTNAME = "hostname";
-    private final FhirContext ctx;
 
-    public OrgRegistrationCommand() {
+    public OrganizationRegistration() {
         super("register", "Register Organization");
-        this.ctx = FhirContext.forDstu3();
-        // Disable server validation, since the Attribution Service doesn't have a capabilities statement
-        this.ctx.getRestfulClientFactory().setServerValidationMode(ServerValidationModeEnum.NEVER);
     }
 
     @Override
-    public void configure(Subparser subparser) {
+    public void addAdditionalOptions(Subparser subparser) {
         // Location of FHIR organization file
         subparser
                 .addArgument("-f", "--file")
                 .dest(ORG_FILE)
                 .type(String.class)
                 .help("FHIR Organization resource to register with system");
-
-        // Address of the Attribution Service, which handles organization registration
-        subparser
-                .addArgument("--host")
-                .dest(ATTR_HOSTNAME)
-                .setDefault("http://localhost:3500/v1")
-                .help("Address of the Attribution Service, which handles organization registration");
     }
 
     @Override
@@ -74,7 +56,7 @@ public class OrgRegistrationCommand extends Command {
         registerOrganization(organization, namespace.getString(ATTR_HOSTNAME));
     }
 
-    void registerOrganization(Bundle organization, String attributionService) throws IOException {
+    private void registerOrganization(Bundle organization, String attributionService) throws IOException {
         final IGenericClient client = ctx.newRestfulGenericClient(attributionService);
 
         final Parameters parameters = new Parameters();
